+++ hapi-fhir-base/src/main/java/ca/uhn/fhir/util/SearchParameterUtil.java
@@ -110,6 +110,15 @@ public class SearchParameterUtil {
 	}
 
 
+	/**
+	 * Return true if any search parameter in the resource can point at a patient, false otherwise
+	 */
+	public static boolean isResourceTypeInPatientCompartment(FhirContext theFhirContext, String theResourceType) {
+		RuntimeResourceDefinition runtimeResourceDefinition = theFhirContext.getResourceDefinition(theResourceType);
+		 return getAllPatientCompartmentRuntimeSearchParams(runtimeResourceDefinition).size() > 0;
+	}
+
+
 	@Nullable
 	public static String getCode(FhirContext theContext, IBaseResource theResource) {
 		return getStringChild(theContext, theResource, "code");
