+++ hapi-fhir-jpaserver-base/src/main/java/ca/uhn/fhir/jpa/bulk/export/svc/BulkDataExportSvcImpl.java
@@ -20,32 +20,20 @@ package ca.uhn.fhir.jpa.bulk.export.svc;
  * #L%
  */
 
-import ca.uhn.fhir.i18n.Msg;
 import ca.uhn.fhir.context.FhirContext;
-import ca.uhn.fhir.context.RuntimeResourceDefinition;
-import ca.uhn.fhir.context.RuntimeSearchParam;
+import ca.uhn.fhir.i18n.Msg;
 import ca.uhn.fhir.interceptor.api.HookParams;
 import ca.uhn.fhir.interceptor.api.IInterceptorBroadcaster;
 import ca.uhn.fhir.interceptor.api.Pointcut;
-import ca.uhn.fhir.jpa.api.config.DaoConfig;
 import ca.uhn.fhir.jpa.api.dao.DaoRegistry;
-import ca.uhn.fhir.jpa.api.dao.IFhirResourceDao;
-import ca.uhn.fhir.jpa.api.model.ExpungeOptions;
-import ca.uhn.fhir.jpa.batch.api.IBatchJobSubmitter;
-import ca.uhn.fhir.jpa.batch.config.BatchConstants;
 import ca.uhn.fhir.jpa.bulk.export.api.IBulkDataExportSvc;
 import ca.uhn.fhir.jpa.bulk.export.model.BulkExportJobStatusEnum;
 import ca.uhn.fhir.jpa.dao.data.IBulkExportCollectionDao;
-import ca.uhn.fhir.jpa.dao.data.IBulkExportCollectionFileDao;
 import ca.uhn.fhir.jpa.dao.data.IBulkExportJobDao;
 import ca.uhn.fhir.jpa.entity.BulkExportCollectionEntity;
 import ca.uhn.fhir.jpa.entity.BulkExportCollectionFileEntity;
 import ca.uhn.fhir.jpa.entity.BulkExportJobEntity;
-import ca.uhn.fhir.jpa.model.sched.HapiJob;
-import ca.uhn.fhir.jpa.model.sched.ISchedulerService;
-import ca.uhn.fhir.jpa.model.sched.ScheduledJobDefinition;
 import ca.uhn.fhir.jpa.model.util.JpaConstants;
-import ca.uhn.fhir.jpa.partition.SystemRequestDetails;
 import ca.uhn.fhir.rest.api.Constants;
 import ca.uhn.fhir.rest.api.server.RequestDetails;
 import ca.uhn.fhir.rest.api.server.bulk.BulkDataExportOptions;
@@ -53,30 +41,19 @@ import ca.uhn.fhir.rest.server.exceptions.InvalidRequestException;
 import ca.uhn.fhir.rest.server.exceptions.ResourceNotFoundException;
 import ca.uhn.fhir.rest.server.servlet.ServletRequestDetails;
 import ca.uhn.fhir.rest.server.util.CompositeInterceptorBroadcaster;
-import ca.uhn.fhir.util.UrlUtil;
+import ca.uhn.fhir.util.SearchParameterUtil;
 import org.apache.commons.lang3.time.DateUtils;
-import org.hl7.fhir.instance.model.api.IBaseBinary;
 import org.hl7.fhir.instance.model.api.IIdType;
 import org.hl7.fhir.r4.model.InstantType;
-import org.quartz.JobExecutionContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.batch.core.JobParametersBuilder;
-import org.springframework.batch.core.JobParametersInvalidException;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
-import org.springframework.transaction.PlatformTransactionManager;
-import org.springframework.transaction.support.TransactionTemplate;
 
-import javax.annotation.PostConstruct;
 import javax.transaction.Transactional;
 import java.util.Date;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
 import java.util.Set;
 import java.util.UUID;
 import java.util.stream.Collectors;
@@ -90,7 +67,6 @@ import static org.apache.commons.lang3.StringUtils.isNotBlank;
 
 public class BulkDataExportSvcImpl implements IBulkDataExportSvc {
 
-	private static final Long READ_CHUNK_SIZE = 10L;
 	private static final Logger ourLog = LoggerFactory.getLogger(BulkDataExportSvcImpl.class);
 	private final int myReuseBulkExportForMillis = (int) (60 * DateUtils.MILLIS_PER_MINUTE);
 
@@ -99,204 +75,13 @@ public class BulkDataExportSvcImpl implements IBulkDataExportSvc {
 	@Autowired
 	private IBulkExportCollectionDao myBulkExportCollectionDao;
 	@Autowired
-	private IBulkExportCollectionFileDao myBulkExportCollectionFileDao;
-	@Autowired
-	private ISchedulerService mySchedulerService;
-	@Autowired
 	private DaoRegistry myDaoRegistry;
 	@Autowired
 	private FhirContext myContext;
-	@Autowired
-	private PlatformTransactionManager myTxManager;
-	private TransactionTemplate myTxTemplate;
-
-	@Autowired
-	private IBatchJobSubmitter myJobSubmitter;
-
-	@Autowired
-	@Qualifier(BatchConstants.BULK_EXPORT_JOB_NAME)
-	private org.springframework.batch.core.Job myBulkExportJob;
-
-	@Autowired
-	@Qualifier(BatchConstants.GROUP_BULK_EXPORT_JOB_NAME)
-	private org.springframework.batch.core.Job myGroupBulkExportJob;
-
-	@Autowired
-	@Qualifier(BatchConstants.PATIENT_BULK_EXPORT_JOB_NAME)
-	private org.springframework.batch.core.Job myPatientBulkExportJob;
 
 	private Set<String> myCompartmentResources;
 
 	private final int myRetentionPeriod = (int) (2 * DateUtils.MILLIS_PER_HOUR);
-
-	/**
-	 * This method is called by the scheduler to run a pass of the
-	 * generator
-	 */
-	@Transactional(value = Transactional.TxType.NEVER)
-	@Override
-	public synchronized void buildExportFiles() {
-		if (!myDaoConfig.isEnableTaskBulkExportJobExecution()) {
-			return;
-		}
-
-		Optional<BulkExportJobEntity> jobToProcessOpt = myTxTemplate.execute(t -> {
-			Pageable page = PageRequest.of(0, 1);
-			Slice<BulkExportJobEntity> submittedJobs = myBulkExportJobDao.findByStatus(page, BulkExportJobStatusEnum.SUBMITTED);
-			if (submittedJobs.isEmpty()) {
-				return Optional.empty();
-			}
-			return Optional.of(submittedJobs.getContent().get(0));
-		});
-
-		if (!jobToProcessOpt.isPresent()) {
-			return;
-		}
-
-		BulkExportJobEntity bulkExportJobEntity = jobToProcessOpt.get();
-
-		String jobUuid = bulkExportJobEntity.getJobId();
-		try {
-				processJob(bulkExportJobEntity);
-		} catch (Exception e) {
-			ourLog.error("Failure while preparing bulk export extract", e);
-			myTxTemplate.execute(t -> {
-				Optional<BulkExportJobEntity> submittedJobs = myBulkExportJobDao.findByJobId(jobUuid);
-				if (submittedJobs.isPresent()) {
-					BulkExportJobEntity jobEntity = submittedJobs.get();
-					jobEntity.setStatus(BulkExportJobStatusEnum.ERROR);
-					jobEntity.setStatusMessage(e.getMessage());
-					myBulkExportJobDao.save(jobEntity);
-				}
-				return null;
-			});
-		}
-
-	}
-
-	private String getQueryParameterIfPresent(String theRequestString, String theParameter) {
-		Map<String, String[]> stringMap = UrlUtil.parseQueryString(theRequestString);
-		if (stringMap != null) {
-			String[] strings = stringMap.get(theParameter);
-			if (strings != null) {
-				return String.join(",", strings);
-			}
-		}
-		return null;
-
-	}
-
-	@Autowired
-	private DaoConfig myDaoConfig;
-
-	/**
-	 * This method is called by the scheduler to run a pass of the
-	 * generator
-	 */
-	@Transactional(value = Transactional.TxType.NEVER)
-	@Override
-	public void purgeExpiredFiles() {
-		if (!myDaoConfig.isEnableTaskBulkExportJobExecution()) {
-			return;
-		}
-
-		Optional<BulkExportJobEntity> jobToDelete = myTxTemplate.execute(t -> {
-			Pageable page = PageRequest.of(0, 1);
-			Slice<BulkExportJobEntity> submittedJobs = myBulkExportJobDao.findByExpiry(page, new Date());
-			if (submittedJobs.isEmpty()) {
-				return Optional.empty();
-			}
-			return Optional.of(submittedJobs.getContent().get(0));
-		});
-
-		if (jobToDelete.isPresent()) {
-
-			ourLog.info("Deleting bulk export job: {}", jobToDelete.get());
-
-			myTxTemplate.execute(t -> {
-
-				BulkExportJobEntity job = myBulkExportJobDao.getOne(jobToDelete.get().getId());
-				for (BulkExportCollectionEntity nextCollection : job.getCollections()) {
-					for (BulkExportCollectionFileEntity nextFile : nextCollection.getFiles()) {
-
-						ourLog.info("Purging bulk data file: {}", nextFile.getResourceId());
-						getBinaryDao().delete(toId(nextFile.getResourceId()), new SystemRequestDetails());
-						getBinaryDao().forceExpungeInExistingTransaction(toId(nextFile.getResourceId()), new ExpungeOptions().setExpungeDeletedResources(true).setExpungeOldVersions(true), new SystemRequestDetails());
-						myBulkExportCollectionFileDao.deleteByPid(nextFile.getId());
-
-					}
-
-					myBulkExportCollectionDao.deleteByPid(nextCollection.getId());
-				}
-
-				ourLog.info("*** ABOUT TO DELETE");
-				myBulkExportJobDao.deleteByPid(job.getId());
-				return null;
-			});
-
-			ourLog.info("Finished deleting bulk export job: {}", jobToDelete.get());
-
-		}
-
-	}
-
-	private void processJob(BulkExportJobEntity theBulkExportJobEntity) {
-		String theJobUuid = theBulkExportJobEntity.getJobId();
-		JobParametersBuilder parameters = new JobParametersBuilder()
-			.addString(BatchConstants.JOB_UUID_PARAMETER, theJobUuid)
-			.addLong(BatchConstants.READ_CHUNK_PARAMETER, READ_CHUNK_SIZE);
-
-		ourLog.info("Submitting bulk export job {} to job scheduler", theJobUuid);
-
-		try {
-			if (isGroupBulkJob(theBulkExportJobEntity)) {
-				enhanceBulkParametersWithGroupParameters(theBulkExportJobEntity, parameters);
-				myJobSubmitter.runJob(myGroupBulkExportJob, parameters.toJobParameters());
-			} else if (isPatientBulkJob(theBulkExportJobEntity)) {
-				myJobSubmitter.runJob(myPatientBulkExportJob, parameters.toJobParameters());
-			} else {
-				myJobSubmitter.runJob(myBulkExportJob, parameters.toJobParameters());
-			}
-		} catch (JobParametersInvalidException theE) {
-			ourLog.error("Unable to start job with UUID: {}, the parameters are invalid. {}", theJobUuid, theE.getMessage());
-		}
-	}
-
-	private boolean isPatientBulkJob(BulkExportJobEntity theBulkExportJobEntity) {
-		return theBulkExportJobEntity.getRequest().startsWith("/Patient/");
-	}
-
-	private boolean isGroupBulkJob(BulkExportJobEntity theBulkExportJobEntity) {
-		return theBulkExportJobEntity.getRequest().startsWith("/Group/");
-	}
-
-	private void enhanceBulkParametersWithGroupParameters(BulkExportJobEntity theBulkExportJobEntity, JobParametersBuilder theParameters) {
-		String theGroupId = getQueryParameterIfPresent(theBulkExportJobEntity.getRequest(), JpaConstants.PARAM_EXPORT_GROUP_ID);
-		String expandMdm  = getQueryParameterIfPresent(theBulkExportJobEntity.getRequest(), JpaConstants.PARAM_EXPORT_MDM);
-		theParameters.addString(BatchConstants.GROUP_ID_PARAMETER, theGroupId);
-		theParameters.addString(BatchConstants.EXPAND_MDM_PARAMETER, expandMdm);
-	}
-
-
-	@SuppressWarnings("unchecked")
-	private IFhirResourceDao<IBaseBinary> getBinaryDao() {
-		return myDaoRegistry.getResourceDao("Binary");
-	}
-
-	@PostConstruct
-	public void start() {
-		myTxTemplate = new TransactionTemplate(myTxManager);
-
-		ScheduledJobDefinition jobDetail = new ScheduledJobDefinition();
-		jobDetail.setId(Job.class.getName());
-		jobDetail.setJobClass(Job.class);
-		mySchedulerService.scheduleClusteredJob(10 * DateUtils.MILLIS_PER_SECOND, jobDetail);
-
-		jobDetail = new ScheduledJobDefinition();
-		jobDetail.setId(PurgeExpiredFilesJob.class.getName());
-		jobDetail.setJobClass(PurgeExpiredFilesJob.class);
-		mySchedulerService.scheduleClusteredJob(DateUtils.MILLIS_PER_HOUR, jobDetail);
-	}
 
 	@Transactional
 	@Override
@@ -391,32 +176,32 @@ public class BulkDataExportSvcImpl implements IBulkDataExportSvc {
 				.filter(t -> !"Binary".equals(t))
 				.collect(Collectors.toSet());
 
-		BulkExportJobEntity job = new BulkExportJobEntity();
-		job.setJobId(UUID.randomUUID().toString());
-		job.setStatus(BulkExportJobStatusEnum.SUBMITTED);
-		job.setSince(since);
-		job.setCreated(new Date());
-		job.setRequest(request);
+		BulkExportJobEntity jobEntity = new BulkExportJobEntity();
+		jobEntity.setJobId(UUID.randomUUID().toString());
+		jobEntity.setStatus(BulkExportJobStatusEnum.SUBMITTED);
+		jobEntity.setSince(since);
+		jobEntity.setCreated(new Date());
+		jobEntity.setRequest(request);
 
 		// Validate types
 		validateTypes(resourceTypes);
 		validateTypeFilters(theBulkDataExportOptions.getFilters(), resourceTypes);
 
-		updateExpiry(job);
-		myBulkExportJobDao.save(job);
+		updateExpiry(jobEntity);
+		myBulkExportJobDao.save(jobEntity);
 
 		for (String nextType : resourceTypes) {
 
 			BulkExportCollectionEntity collection = new BulkExportCollectionEntity();
-			collection.setJob(job);
+			collection.setJob(jobEntity);
 			collection.setResourceType(nextType);
-			job.getCollections().add(collection);
+			jobEntity.getCollections().add(collection);
 			myBulkExportCollectionDao.save(collection);
 		}
 
-		ourLog.info("Bulk export job submitted: {}", job.toString());
+		ourLog.info("Bulk export job submitted: {}", jobEntity.toString());
 
-		return toSubmittedJobInfo(job);
+		return toSubmittedJobInfo(jobEntity);
 	}
 
 	public void validateTypes(Set<String> theResourceTypes) {
@@ -484,19 +269,10 @@ public class BulkDataExportSvcImpl implements IBulkDataExportSvc {
 	public Set<String> getPatientCompartmentResources() {
 		if (myCompartmentResources == null) {
 			myCompartmentResources = myContext.getResourceTypes().stream()
-				.filter(this::resourceTypeIsInPatientCompartment)
+				.filter(resType -> SearchParameterUtil.isResourceTypeInPatientCompartment(myContext, resType))
 				.collect(Collectors.toSet());
 		}
 		return myCompartmentResources;
-	}
-
-	/**
-	 * Return true if any search parameter in the resource can point at a patient, false otherwise
-	 */
-	private boolean resourceTypeIsInPatientCompartment(String theResourceType) {
-		RuntimeResourceDefinition runtimeResourceDefinition = myContext.getResourceDefinition(theResourceType);
-		List<RuntimeSearchParam> searchParams = runtimeResourceDefinition.getSearchParamsForCompartmentName("Patient");
-		return searchParams != null && searchParams.size() >= 1;
 	}
 
 	public Set<String> getAllowedResourceTypesForBulkExportStyle(BulkDataExportOptions.ExportStyle theExportStyle) {
@@ -509,11 +285,6 @@ public class BulkDataExportSvcImpl implements IBulkDataExportSvc {
 		}
 	}
 
-	private IIdType toId(String theResourceId) {
-		IIdType retVal = myContext.getVersion().newIdType();
-		retVal.setValue(theResourceId);
-		return retVal;
-	}
 
 	private IIdType toQualifiedBinaryId(String theIdPart) {
 		IIdType retVal = myContext.getVersion().newIdType();
@@ -522,38 +293,5 @@ public class BulkDataExportSvcImpl implements IBulkDataExportSvc {
 		return retVal;
 	}
 
-	@Override
-	@Transactional(Transactional.TxType.NEVER)
-	public synchronized void cancelAndPurgeAllJobs() {
-		myTxTemplate.execute(t -> {
-			ourLog.info("Deleting all files");
-			myBulkExportCollectionFileDao.deleteAllFiles();
-			ourLog.info("Deleting all collections");
-			myBulkExportCollectionDao.deleteAllFiles();
-			ourLog.info("Deleting all jobs");
-			myBulkExportJobDao.deleteAllFiles();
-			return null;
-		});
-	}
-
-	public static class Job implements HapiJob {
-		@Autowired
-		private IBulkDataExportSvc myTarget;
-
-		@Override
-		public void execute(JobExecutionContext theContext) {
-			myTarget.buildExportFiles();
-		}
-	}
-
-	public static class PurgeExpiredFilesJob implements HapiJob {
-		@Autowired
-		private IBulkDataExportSvc myTarget;
-
-		@Override
-		public void execute(JobExecutionContext theContext) {
-			myTarget.purgeExpiredFiles();
-		}
-	}
 
 }
