+++ hapi-fhir-storage/src/main/java/ca/uhn/fhir/jpa/bulk/export/api/IBulkDataExportJobSchedulingHelper.java
@@ -1,0 +1,11 @@
+package ca.uhn.fhir.jpa.bulk.export.api;
+
+import javax.transaction.Transactional;
+
+public interface IBulkDataExportJobSchedulingHelper {
+
+	@Transactional(value = Transactional.TxType.NEVER)
+	void purgeExpiredFiles();
+	void cancelAndPurgeAllJobs();
+	void startSubmittedJobs();
+}
